+++ node/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pylo/node
 
+## 0.2.0
+
+### Minor Changes
+
+- adds upsert functionality to sdks
+
 ## 0.1.1
 
 ### Patch Changes